+++ SM_Mono/SM_Component.cs
@@ -18,8 +18,8 @@ namespace MelonLoader.Support
         }
         private void SiblingFix() { SetAsLastSiblingMethod?.Invoke(gameObject.transform, new object[0]); SetAsLastSiblingMethod?.Invoke(transform, new object[0]); }
         internal void Destroy() { IsDestroying = true; GameObject.Destroy(gameObject); }
+        void Start() { SiblingFix(); Main.Interface.OnApplicationStart(); }
         void Awake() { foreach (var queuedCoroutine in SupportModule_To.QueuedCoroutines) StartCoroutine(queuedCoroutine); SupportModule_To.QueuedCoroutines.Clear(); }
-        void Start() => SiblingFix();
         void Update() { SiblingFix(); Main.Interface.Update(); }
         void FixedUpdate() => Main.Interface.FixedUpdate();
         void LateUpdate() => Main.Interface.LateUpdate();